requirements.txt
# Dependencias del proyecto POS HTF
PySide6>=6.6.0  # Framework de UI moderno con Qt6
QtAwesome>=1.4.0  # Iconos FontAwesome para Qt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # Driver PostgreSQL
bcrypt>=4.0.0  # Para hashear contraseñas
requests>=2.31.0  # HTTP requests
pytz>=2023.3  # Zonas horarias para Python
pyinstaller>=6.0.0  # Para generar ejecutable de Windows
customtkinter>=5.2.2  # UI alternativa para formularios simples
openpyxl>=3.11.0  # Leer y escribir archivos Excel (.xlsx)
pandas>=2.0.0  # Manipulación de datos (opcional pero útil)
xlsxwriter>=3.1.2  # Alternativa para crear Excel con formato avanzado